Add files via upload
modified serverevents and added new env variable for serverdisplay_name to use in serverevents to display as different servers, currently setup to read bitquest and use the default bits denomination as what the server runs on. eventually. i would like to cahnge defaults to everything to an emerald server basically and add the crypto if wanted... idk.. 

also.... cant seem to get claimland in bitquest.java to work with emeralds? maybe @explodi can look at please?